created the reset password and forgot password route

+++ controller/butcamps.js
@@ -36,7 +36,16 @@ exports.createBootcamps = asyncHandler(async (req, res, next) => {
   // Add user to req.body
   req.body.user = req.user.id
 
-  // console.log(req.body);
+  // check for published bootcamp
+  const publishedBootcamp = await BootcampModel.findOne({user: req.user.id})
+  
+
+// If the user is not an admin, they can only publish one bootcamp
+if(publishedBootcamp && req.user.role !== 'admin'){
+  return next(
+    new ErrorResponse(`User with ID ${req.user.id} already published a bootcamp`, 400)
+  );
+}
 
   const bootcamp = await BootcampModel.create(req.body);
 
@@ -51,19 +60,29 @@ exports.createBootcamps = asyncHandler(async (req, res, next) => {
 // @route      Put/api/v1/bootcamps/:id
 // @access     Private
 exports.updateBootcamp = asyncHandler(async (req, res, next) => {
-  const upBoot = await BootcampModel.findByIdAndUpdate(
+  let upBoot = await BootcampModel.findById(req.params.id);
+  if (!upBoot) {
+    return next(
+      new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
+    );
+  }
+
+  // Make sure user is the bootcamp owner 
+  if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin'){
+    return next(
+      new ErrorResponse(`User with ID of ${req.user.id} and name ${req.user.name} is not authorized to update this bootcamp`, 401)
+    );
+  }
+
+  upBoot = await BootcampModel.findByIdAndUpdate(
     req.params.id,
     req.body,
     {
       new: true,
       runValidators: true
     }
-  );
-  if (!upBoot) {
-    return next(
-      new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
-    );
-  }
+  )
+
   res.status(200).json({
     success: true,
     data: upBoot
@@ -82,6 +101,12 @@ exports.deleteBootcamp = asyncHandler(async (req, res, next) => {
     );
   }
 
+ // Make sure user is the bootcamp owner 
+ if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin'){
+  return next(
+    new ErrorResponse(`User with ID of ${req.user.id} and name ${req.user.name} is not authorized to delete this bootcamp`, 401)
+  );
+}
 
  await bootcamp.deleteOne() 
 
@@ -127,6 +152,14 @@ exports.BootcampPhoto = asyncHandler(async (req, res, next) => {
       new ErrorResponse(`Please upload an image less than ${process.env.MAX_FILE_UPLOAD}`, 400)
     );
   }
+
+   // Make sure user is the bootcamp owner 
+   if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin'){
+    return next(
+      new ErrorResponse(`User with ID of ${req.user.id} and name ${req.user.name} is not authorized to upload photo to this bootcamp`, 401)
+    );
+  }
+
 
   file.name = `photo_${bootcamp._id}${path.parse(file.name).ext}`
